Add label data visibility, translation options and some enhancements
Added an option to choose to show in the labels different data, such as the percentages (like before), the data itself, or both of them.
Added an options to choose if you want to use the Godot Translation Server in the labels, titles, etc.
Added data prefix/suffix when showed.
Improved label formatting as well as refactors the label rendering logic.
Now if you call the graph to render without any data, it will show an empty graph instead of no graph at all.
Finally, added a function to receive a new data dictionary.